+++ src/main/java/ru/complitex/address/entity/District.java
@@ -7,12 +7,12 @@ import ru.complitex.domain.entity.NameDomain;
  * 02.04.2020 9:38 PM
  */
 public class District extends NameDomain<District> {
-    public final static Long ENTITY_ID = 5L;
+    public final static int ENTITY_ID = 5;
     public final static String ENTITY_NAME = "district";
 
-    public final static Long CITY = 1L;
-    public final static Long NAME = 2L;
-    public final static Long CODE = 3L;
+    public final static int CITY = 1;
+    public final static int NAME = 2;
+    public final static int CODE = 3;
 
     public District() {
         super(ENTITY_ID, ENTITY_NAME, NAME);
