+++ src/main/java/ru/complitex/address/entity/Room.java
@@ -7,11 +7,11 @@ import ru.complitex.domain.entity.Domain;
  * 27.04.2020 22:33
  */
 public class Room extends Domain<Room> {
-    public final static Long ENTITY_ID = 10L;
+    public final static int ENTITY_ID = 10;
     public final static String ENTITY_NAME = "room";
 
-    public final static Long BUILDING = 1L;
-    public final static Long NAME = 2L;
+    public final static int BUILDING = 1;
+    public final static int NAME = 2;
 
     public Room() {
         super(ENTITY_ID, ENTITY_NAME);
