+++ src/main/java/ru/complitex/address/entity/Street.java
@@ -7,13 +7,13 @@ import ru.complitex.domain.entity.NameDomain;
  * 02.04.2020 9:42 PM
  */
 public class Street extends NameDomain<Street> {
-    public final static Long ENTITY_ID = 7L;
+    public final static int ENTITY_ID = 7;
     public final static String ENTITY_NAME = "street";
 
-    public final static Long CITY = 1L;
-    public final static Long STREET_TYPE = 2L;
-    public final static Long NAME = 3L;
-    public final static Long CODE = 4L;
+    public final static int CITY = 1;
+    public final static int STREET_TYPE = 2;
+    public final static int NAME = 3;
+    public final static int CODE = 4;
 
     public Street() {
         super(ENTITY_ID, ENTITY_NAME, NAME);
