+++ src/main/java/ru/complitex/address/entity/StreetType.java
@@ -7,11 +7,11 @@ import ru.complitex.domain.entity.ShortNameDomain;
  * 02.04.2020 9:41 PM
  */
 public class StreetType extends ShortNameDomain<StreetType> {
-    public final static Long ENTITY_ID = 6L;
+    public final static int ENTITY_ID = 6;
     public final static String ENTITY_NAME = "street_type";
 
-    public final static Long NAME = 1L;
-    public final static Long SHORT_NAME = 2L;
+    public final static int NAME = 1;
+    public final static int SHORT_NAME = 2;
 
     public StreetType() {
         super(ENTITY_ID, ENTITY_NAME, NAME, SHORT_NAME);
